#258 [refactor] 총 글감 수 조회와 페이지 내 글감 리스트 조회 분리

+++ module-api/src/main/java/com/mile/controller/moim/MoimController.java
@@ -153,7 +153,7 @@ public class MoimController implements MoimControllerSwagger {
     @GetMapping("/{moimId}/admin/topicList")
     public ResponseEntity<SuccessResponse<MoimTopicInfoListResponse>> getMoimTopicList(
             @MoimIdPathVariable final Long moimId,
-            @RequestParam final Long page,
+            @RequestParam final int page,
             @PathVariable("moimId") final String moimUrl
     ) {
         return ResponseEntity.ok(SuccessResponse.of(SuccessMessage.MOIM_TOPIC_LIST_GET_SUCCESS, moimService.getMoimTopicList(moimId, principalHandler.getUserIdFromPrincipal(), page)));
